Moving paparazzi.
