+++ app/src/main/java/com/adammcneilly/spacenerd/screens/astronauts/AstronautListScreen.kt
@@ -23,7 +23,7 @@ fun AstronautListScreen(
             TopAppBar(
                 title = {
                     Text(
-                        text = "Astronuts",
+                        text = "Astronauts",
                     )
                 },
             )
