+++ shared/src/test/kotlin/com/adammcneilly/spacenerd/shared/feature/launchlist/ui/LaunchSummaryCardPaparazziTest.kt
@@ -1,7 +1,6 @@
-package com.adammcneilly.spacenerd.screens.launches
+package com.adammcneilly.spacenerd.shared.feature.launchlist.ui
 
-import com.adammcneilly.spacenerd.BasePaparazziTest
-import com.adammcneilly.spacenerd.R
+import com.adammcneilly.spacenerd.shared.BasePaparazziTest
 import com.adammcneilly.spacenerd.shared.core.displaymodels.ImageDisplayModel
 import com.adammcneilly.spacenerd.shared.core.displaymodels.LaunchDisplayModel
 import com.adammcneilly.spacenerd.shared.core.displaymodels.LaunchStatusDisplayModel
@@ -25,7 +24,7 @@ class LaunchSummaryCardPaparazziTest : BasePaparazziTest() {
         val launch = LaunchDisplayModel(
             id = "123",
             name = "Falcon 9 Block 5 | Starlink Group 12-19",
-            image = ImageDisplayModel.AndroidLocal(R.drawable.falcon9),
+            image = ImageDisplayModel.Placeholder,
             status = LaunchStatusDisplayModel(LaunchStatus.Go),
             subtitle = "SpaceX • Cape Canaveral SFS, FL, USA",
         )
