+++ shared/src/test/snapshots/images/com.adammcneilly.spacenerd.shared.feature.launchlist.ui_LaunchSummaryCardPaparazziTest_renderPlaceholder[LANDSCAPE_PHONE].png
