+++ shared/src/test/snapshots/images/com.adammcneilly.spacenerd.shared.feature.launchlist.ui_LaunchSummaryCardPaparazziTest_renderPlaceholder[LIGHT_PHONE].png
